+++ go-app/config/config_profile.go
@@ -131,6 +131,8 @@ type Config_Controller_Profile_Control struct {
 type Config_Controller_Profile_Controller struct {
 	/* if defined ; specifies this profile can only be used with the below controller */
 	UsbID *string `json:"usb_id,omitempty"`
+	/* specifies if this profile can be autoselected for this controller */
+	AutoSelect *bool `json:"auto_select"`
 	/* Can be defined to specify a specific SDL mapping for this controller and profile; useful for sharing */
 	Mapping *Config_Controller_SDLMap `json:"mapping,omitempty"`
 }