+++ go-app/frontend/src/store.ts
@@ -1,7 +1,0 @@
-import { create } from 'zustand'
-
-type StoreType = {
-}
-
-export const useStore = create<StoreType>((set) => ({
-}))