+++ go-app/windows_build.ps1
@@ -1,3 +1,4 @@
+Write-Host "Building version: $env:VERSION"
 $env:CGO_CFLAGS="-I$PWD/vcpkg_installed/x64-windows/include"
 $env:CGO_LDFLAGS="-L$PWD/vcpkg_installed/x64-windows/lib -lSDL2"
-wails build ldflags="-X 'main.VERSION=$VERSION'"
+wails build ldflags="-X 'main.VERSION=$env:VERSION'"